feat: use ui-tab-routes in dev route
