+++ packages/dev/src/dev-feature-ui.tsx
@@ -1,0 +1,11 @@
+import { DevFeatureUiAvatars } from './dev-feature-ui-avatars.js'
+import { DevFeatureUiColors } from './dev-feature-ui-colors.js'
+
+export default function DevFeatureUi() {
+  return (
+    <div className="space-y-6">
+      <DevFeatureUiAvatars />
+      <DevFeatureUiColors />
+    </div>
+  )
+}
